snowflake/pyproject.toml
[tool.poetry]
name = "dbt"
version = "0.1.0"
description = ""
authors = ["raphelemmanuvel"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
dbt-snowflake = "^1.7.0"
dbt-core = "^1.7.2"
typer = "^0.9.0"
rich = "^13.7.0"
snowflake-connector-python = "^3.5.0"
black = "^23.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
